+++ src/components/ListTemplate/actionTypes.js
@@ -4,20 +4,15 @@ export const ACTION_TYPES = {
     iconKey: 'plus-sign',
     captionId: 'ui-data-import.create',
   },
-  COMBINE: {
-    type: 'COMBINE',
-    iconKey: 'combine',
-    captionId: 'ui-data-import.combine',
-  },
   MODIFY: {
     type: 'MODIFY',
     iconKey: 'edit',
     captionId: 'ui-data-import.modify',
   },
-  REPLACE: {
-    type: 'REPLACE',
+  UPDATE: {
+    type: 'UPDATE',
     iconKey: 'replace',
-    captionId: 'ui-data-import.replace',
+    captionId: 'ui-data-import.update',
   },
 };
 
@@ -26,16 +21,12 @@ export const ACTION_TYPES_SELECT = {
     ...ACTION_TYPES.CREATE,
     captionId: 'ui-data-import.actionSelect.create',
   },
-  COMBINE: {
-    ...ACTION_TYPES.COMBINE,
-    captionId: 'ui-data-import.actionSelect.combine',
-  },
   MODIFY: {
     ...ACTION_TYPES.MODIFY,
     captionId: 'ui-data-import.actionSelect.modify',
   },
-  REPLACE: {
-    ...ACTION_TYPES.REPLACE,
-    captionId: 'ui-data-import.actionSelect.replace',
+  UPDATE: {
+    ...ACTION_TYPES.UPDATE,
+    captionId: 'ui-data-import.actionSelect.update',
   },
 };