+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Update MARC matching UI to differentiate [any] versus [blank] indicators (UIDATIMP-525)
 * Reuse `<FullScreenForm>` component from `stripes-data-transfer-components` repository (UIDATIMP-578)
 * Add option for Modify or Update MARC Bib field mapping profile (UIDATIMP-612)
+* Update the available actions for Action profiles (UIDATIMP-559)
 
 ### Bugs fixed:
 * Fix rendering qualifier sections with old data in match profiles details (UIDATIMP-481)